requirements.txt
Flask
python-telegram-bot
PyPDF2
docxtpl
requests
gunicorn
